backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pillow
pytesseract
google-generativeai
python-dotenv
python-magic-bin; platform_system == "Windows"
python-magic; platform_system != "Windows"
